Refactor Results page to ResultsContainer and create Results component
