+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom'
 // Pages
 import Home from './pages/Home/Home'
-import Results from './pages/Results/Results'
+import ResultsContainer from './pages/ResultsContainer/ResultsContainer'
 import DetailContainer from './pages/DetailContainer/DetailContainer'
 import NotFound from './pages/NotFound/NotFound'
 
@@ -10,7 +10,7 @@ function App() {
     <>
       <Routes>
         <Route path='/' element={<Home />} />
-        <Route path='/items' element={<Results />} />
+        <Route path='/items' element={<ResultsContainer />} />
         <Route path='/items/:id' element={<DetailContainer />} />
         <Route path='*' element={<NotFound />} />
       </Routes>
