--- src/pages/Results/results.css
@@ -1,34 +1,0 @@
-main.results{
-  padding-top: 0;
-}
-
-.results h1 {
-  font-size: 1em;
-}
-
-.results ul {
-  display: flex;
-  flex-direction: column;
-}
-
-li.card {
-  display: flex;
-  gap: 20px;
-  min-height: 150px;
-  align-items: center;
-}
-
-li.card img {
-  width: 100px;
-  height: 100px;
-}
-
-li.card .price-section {
-  display: flex;
-  justify-content: space-between;
-  margin: 5px 0;
-}
-
-.card .description{
-  font-size: .9em;
-}
