+++ src/pages/ResultsContainer/resultsContainer.css
@@ -1,0 +1,3 @@
+main.results{
+  padding-top: 0;
+}
